split Checkout into multiple Components
